jwt에서 memberNo받아오기 위한 수정 작업(최경환, main)
AuthUser.java 추가
JwtAuthenticationFilter.java 추가
memberNo long으로 되어있던거 int로 수정
securityconfig addFilterBefore 추가

+++ BackEnd_TripType_Final_Project/src/main/java/com/kh/triptype/auth/service/AuthService.java
@@ -73,7 +73,7 @@ public class AuthService {
 
         // 5️ JWT 발급
         String token = jwtProvider.createToken(
-                (long) member.getMemberNo(),
+                member.getMemberNo(),
                 member.getMemberRole()
         );
